API/requirements.txt
fastapi
uvicorn
pydantic
mysql-connector-python
python-multipart